- fix warning C4311: 'type cast': pointer truncation from 'fi_handle' to 'int' - fix error C2440: '=': cannot convert from 'BOOL (__cdecl *)(FreeImageIO &,fi_handle)' to 'FI_ValidateProc' - fix error C2440: 'initializing': cannot convert from 'const char [9]' to 'char *' - fix error C2664: 'BOOL testMemoryStreamMultiPageOpenSave(const char *,char *,int,int)': cannot convert argument 2 from 'const char [27]' to 'char *' - fix warning C4777: 'sprintf' : format string '%X' requires an argument of type 'unsigned int', but variadic argument 1 has type 'DWORD' - fix several more C/C++ compiler errors re include files, etc.

+++ Examples/Generic/LoadFromMemory.cpp
@@ -68,9 +68,9 @@ _SeekProc(fi_handle handle, long offset, int origin) {
 
 inline long _stdcall
 _TellProc(fi_handle handle) {
-	assert((int)handle > (int)g_load_address);
+	assert((intptr_t)handle > (intptr_t)g_load_address);
 
-	return ((int)g_load_address - (int)handle);
+	return ((intptr_t)g_load_address - (intptr_t)handle);
 }
 
 // ----------------------------------------------------------
